bacswn/requirements.txt
# BACSWN SkyWatch Bahamas
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1
websockets==14.1
pydantic==2.10.3
apscheduler==3.10.4
